+++ lib/properties.ml
@@ -3364,7 +3364,395 @@ let canonical_initial_for_minify : type a. a property -> a -> a =
   | Height, Length Initial -> Length Auto
   | Min_width, Length Initial -> Length Auto
   | Min_height, Length Initial -> Length Auto
-  | _ -> value
+  | Min_inline_size, Length Initial -> Length Auto
+  | Min_block_size, Length Initial -> Length Auto
+  (* Keep this fallback exhaustive: a new property must make this match fail to
+     compile until its initial-value fold has been considered. *)
+  | Custom_property _, value -> value
+  | Unknown_property _, value -> value
+  | All, value -> value
+  | ( ( Background_color | Color | Text_decoration_color | Text_emphasis_color
+      | Border_top_color | Border_right_color | Border_bottom_color
+      | Border_left_color | Border_inline_start_color | Border_inline_end_color
+      | Border_block_start_color | Border_block_end_color | Outline_color
+      | Webkit_tap_highlight_color | Webkit_text_decoration_color
+      | Webkit_text_fill_color | Webkit_text_stroke_color | Column_rule_color
+      | Stop_color | Flood_color | Lighting_color | Accent_color | Caret_color
+        ),
+      value ) ->
+      value
+  | Border_color, value -> value
+  | ( ( Border_style | Border_top_style | Border_right_style
+      | Border_bottom_style | Border_left_style | Border_inline_start_style
+      | Border_inline_end_style | Border_block_start_style
+      | Border_block_end_style | Border_inline_style | Border_block_style ),
+      value ) ->
+      value
+  | ( ( Padding | Padding_inline | Padding_block | Margin | Margin_inline
+      | Margin_block | Inset | Inset_inline | Inset_inline_start
+      | Inset_inline_end | Inset_block | Inset_block_start | Inset_block_end
+      | Top | Right | Bottom | Left | Scroll_margin | Scroll_margin_inline
+      | Scroll_margin_block | Scroll_padding | Scroll_padding_inline
+      | Scroll_padding_block ),
+      value ) ->
+      value
+  | ( ( Padding_left | Padding_right | Padding_bottom | Padding_top
+      | Padding_inline_start | Padding_inline_end | Padding_block_start
+      | Padding_block_end | Margin_inline_end | Margin_inline_start
+      | Margin_left | Margin_right | Margin_top | Margin_bottom
+      | Margin_block_start | Margin_block_end | Column_gap | Row_gap
+      | Text_underline_offset | Letter_spacing | Border_top_left_radius
+      | Border_top_right_radius | Border_bottom_left_radius
+      | Border_bottom_right_radius | Border_start_start_radius
+      | Border_start_end_radius | Border_end_start_radius
+      | Border_end_end_radius | Outline_offset | Line_height_step | Perspective
+      | Word_spacing | Text_decoration_thickness | Scroll_margin_top
+      | Scroll_margin_right | Scroll_margin_bottom | Scroll_margin_left
+      | Scroll_margin_inline_start | Scroll_margin_inline_end
+      | Scroll_margin_block_start | Scroll_margin_block_end | Scroll_padding_top
+      | Scroll_padding_right | Scroll_padding_bottom | Scroll_padding_left
+      | Scroll_padding_inline_start | Scroll_padding_inline_end
+      | Scroll_padding_block_start | Scroll_padding_block_end ),
+      value ) ->
+      value
+  | Gap, value -> value
+  | ( ( Width | Height | Min_width | Min_height | Max_width | Max_height
+      | Inline_size | Min_inline_size | Max_inline_size | Block_size
+      | Min_block_size | Max_block_size | Shape_margin | Offset_distance ),
+      value ) ->
+      value
+  | Font_size, value -> value
+  | Line_height, value -> value
+  | Font_weight, value -> value
+  | Font_style, value -> value
+  | Text_align, value -> value
+  | (Text_decoration | Webkit_text_decoration), value -> value
+  | Text_decoration_line, value -> value
+  | Text_decoration_style, value -> value
+  | Text_decoration_skip, value -> value
+  | Text_decoration_skip_self, value -> value
+  | Text_decoration_skip_box, value -> value
+  | Text_decoration_skip_inset, value -> value
+  | Text_decoration_skip_spaces, value -> value
+  | Text_emphasis, value -> value
+  | Text_emphasis_style, value -> value
+  | Text_emphasis_position, value -> value
+  | Text_emphasis_skip, value -> value
+  | Text_orientation, value -> value
+  | Text_transform, value -> value
+  | List_style_type, value -> value
+  | List_style_position, value -> value
+  | List_style_image, value -> value
+  | Display, value -> value
+  | Position, value -> value
+  | Visibility, value -> value
+  | Baseline_source, value -> value
+  | Alignment_baseline, value -> value
+  | Baseline_shift, value -> value
+  | (Flex_direction | Webkit_flex_direction), value -> value
+  | (Flex_wrap | Webkit_flex_wrap), value -> value
+  | (Flex_flow | Webkit_flex_flow), value -> value
+  | Flex, value -> value
+  | (Flex_grow | Flex_shrink), value -> value
+  | Flex_basis, value -> value
+  | Order, value -> value
+  | (Align_items | Webkit_align_items), value -> value
+  | (Justify_content | Webkit_justify_content), value -> value
+  | Justify_items, value -> value
+  | Justify_self, value -> value
+  | (Align_content | Webkit_align_content), value -> value
+  | (Align_self | Webkit_align_self), value -> value
+  | Place_content, value -> value
+  | Place_items, value -> value
+  | Place_self, value -> value
+  | ( ( Grid_template_columns | Grid_template_rows | Grid_template | Grid
+      | Grid_auto_columns | Grid_auto_rows ),
+      value ) ->
+      value
+  | Grid_template_areas, value -> value
+  | Grid_area, value -> value
+  | Grid_auto_flow, value -> value
+  | (Grid_column | Grid_row), value -> value
+  | (Grid_column_start | Grid_column_end | Grid_row_start | Grid_row_end), value
+    ->
+      value
+  | Border_width, value -> value
+  | ( ( Border_top_width | Border_right_width | Border_bottom_width
+      | Border_left_width | Border_inline_start_width | Border_inline_end_width
+      | Border_block_start_width | Border_block_end_width | Outline_width ),
+      value ) ->
+      value
+  | (Border_inline_width | Border_block_width), value -> value
+  | (Border_image | Mask_border), value -> value
+  | (Border_image_source | Webkit_mask_image | Mask_image), value -> value
+  | Border_image_slice, value -> value
+  | Border_image_repeat, value -> value
+  | Border_image_width, value -> value
+  | Border_image_outset, value -> value
+  | (Border_radius | Webkit_border_radius | Moz_border_radius), value -> value
+  | (Border_inline_color | Border_block_color), value -> value
+  | ( (Opacity | Fill_opacity | Stroke_opacity | Stop_opacity | Flood_opacity),
+      value ) ->
+      value
+  | Mix_blend_mode, value -> value
+  | ( (Transform | Webkit_transform | Moz_transform | Ms_transform | O_transform),
+      value ) ->
+      value
+  | Translate, value -> value
+  | Cursor, value -> value
+  | Interactivity, value -> value
+  | Caret_animation, value -> value
+  | Caret_shape, value -> value
+  | Caret, value -> value
+  | (Interest_delay | Interest_delay_start | Interest_delay_end), value -> value
+  | (Nav_up | Nav_right | Nav_down | Nav_left), value -> value
+  | Table_layout, value -> value
+  | Border_collapse, value -> value
+  | Border_spacing, value -> value
+  | (User_select | Webkit_user_select | Moz_user_select | Ms_user_select), value
+    ->
+      value
+  | Pointer_events, value -> value
+  | ( (Overflow | Overflow_x | Overflow_y | Overflow_block | Overflow_inline),
+      value ) ->
+      value
+  | Z_index, value -> value
+  | Outline, value -> value
+  | Outline_style, value -> value
+  | Forced_color_adjust, value -> value
+  | Scroll_snap_type, value -> value
+  | White_space, value -> value
+  | ( ( Border | Border_block | Border_block_start | Border_block_end
+      | Border_inline | Border_inline_start | Border_inline_end | Column_rule
+      | Border_top | Border_right | Border_bottom | Border_left ),
+      value ) ->
+      value
+  | Background, value -> value
+  | Tab_size, value -> value
+  | Zoom, value -> value
+  | (Webkit_text_size_adjust | Text_size_adjust), value -> value
+  | Font_feature_settings, value -> value
+  | Font_variation_settings, value -> value
+  | Text_indent, value -> value
+  | List_style, value -> value
+  | Font, value -> value
+  | Source, value -> value
+  | Webkit_appearance, value -> value
+  | (Webkit_transition | Moz_transition | O_transition | Transition), value ->
+      value
+  | ( ( Webkit_transition_delay | Webkit_transition_duration
+      | Webkit_animation_delay | Webkit_animation_duration | Moz_animation_delay
+      | Moz_animation_duration | Moz_transition_delay | Moz_transition_duration
+      | Transition_duration | Transition_delay | Animation_duration
+      | Animation_delay ),
+      value ) ->
+      value
+  | ( ( Webkit_transition_property | Moz_transition_property
+      | Transition_property ),
+      value ) ->
+      value
+  | ( ( Webkit_transition_timing_function | Webkit_animation_timing_function
+      | Moz_animation_timing_function | Moz_transition_timing_function
+      | Transition_timing_function | Animation_timing_function ),
+      value ) ->
+      value
+  | (Webkit_animation | Moz_animation | Animation), value -> value
+  | ( ( Webkit_animation_direction | Moz_animation_direction
+      | Animation_direction ),
+      value ) ->
+      value
+  | ( ( Webkit_animation_iteration_count | Moz_animation_iteration_count
+      | Animation_iteration_count ),
+      value ) ->
+      value
+  | (Webkit_animation_name | Moz_animation_name | Animation_name), value ->
+      value
+  | ( ( Webkit_animation_fill_mode | Moz_animation_fill_mode
+      | Animation_fill_mode ),
+      value ) ->
+      value
+  | ( ( Webkit_animation_play_state | Moz_animation_play_state
+      | Animation_play_state ),
+      value ) ->
+      value
+  | (Webkit_box_sizing | Moz_box_sizing | Box_sizing), value -> value
+  | (Webkit_box_shadow | Moz_box_shadow | Box_shadow), value -> value
+  | ( (Webkit_background_size | Background_size | Webkit_mask_size | Mask_size),
+      value ) ->
+      value
+  | ( ( Webkit_filter | Ms_filter | Filter | Backdrop_filter
+      | Webkit_backdrop_filter ),
+      value ) ->
+      value
+  | (Moz_appearance | Appearance), value -> value
+  | Container_type, value -> value
+  | Container_name, value -> value
+  | Container, value -> value
+  | Anchor_name, value -> value
+  | Position_anchor, value -> value
+  | Position_try_fallbacks, value -> value
+  | Position_try_order, value -> value
+  | Position_try, value -> value
+  | Position_visibility, value -> value
+  | Position_area, value -> value
+  | Shape_outside, value -> value
+  | Shape_image_threshold, value -> value
+  | Overflow_clip_margin, value -> value
+  | Overflow_anchor, value -> value
+  | Scrollbar_width, value -> value
+  | Scrollbar_color, value -> value
+  | Scrollbar_gutter, value -> value
+  | Font_palette, value -> value
+  | Font_synthesis, value -> value
+  | Text_wrap_mode, value -> value
+  | Text_wrap_style, value -> value
+  | Text_box_trim, value -> value
+  | Text_underline_position, value -> value
+  | Text_box_edge, value -> value
+  | Text_box, value -> value
+  | Inline_sizing, value -> value
+  | Line_fit_edge, value -> value
+  | Interpolate_size, value -> value
+  | Min_intrinsic_sizing, value -> value
+  | Ruby_align, value -> value
+  | Ruby_merge, value -> value
+  | Ruby_overhang, value -> value
+  | Ruby_position, value -> value
+  | Glyph_orientation_vertical, value -> value
+  | Text_combine_upright, value -> value
+  | Animation_timeline, value -> value
+  | Animation_range, value -> value
+  | (Animation_range_start | Animation_range_end), value -> value
+  | (Scroll_timeline | View_timeline), value -> value
+  | (Scroll_timeline_name | View_timeline_name | Timeline_scope), value -> value
+  | (Scroll_timeline_axis | View_timeline_axis), value -> value
+  | View_transition_name, value -> value
+  | View_transition_class, value -> value
+  | Image_orientation, value -> value
+  | Image_rendering, value -> value
+  | Image_resolution, value -> value
+  | Contain_intrinsic_size, value -> value
+  | ( ( Contain_intrinsic_width | Contain_intrinsic_height
+      | Contain_intrinsic_block_size | Contain_intrinsic_inline_size ),
+      value ) ->
+      value
+  | Margin_trim, value -> value
+  | Offset_path, value -> value
+  | Offset_rotate, value -> value
+  | Font_size_adjust, value -> value
+  | Font_variant_emoji, value -> value
+  | Text_spacing_trim, value -> value
+  | Hyphenate_limit_chars, value -> value
+  | Initial_letter, value -> value
+  | Initial_letter_align, value -> value
+  | Initial_letter_wrap, value -> value
+  | Dominant_baseline, value -> value
+  | View_timeline_inset, value -> value
+  | Perspective_origin, value -> value
+  | Transform_style, value -> value
+  | Backface_visibility, value -> value
+  | Object_position, value -> value
+  | Rotate, value -> value
+  | Transition_behavior, value -> value
+  | Overlay, value -> value
+  | Will_change, value -> value
+  | Contain, value -> value
+  | Isolation, value -> value
+  | (Break_before | Break_after), value -> value
+  | Break_inside, value -> value
+  | (Page_break_before | Page_break_after), value -> value
+  | Page_break_inside, value -> value
+  | Page_size, value -> value
+  | Columns, value -> value
+  | Column_width, value -> value
+  | Column_count, value -> value
+  | Column_span, value -> value
+  | Background_attachment, value -> value
+  | Transform_origin, value -> value
+  | Transform_box, value -> value
+  | Text_shadow, value -> value
+  | Clip_path, value -> value
+  | Mask, value -> value
+  | Content_visibility, value -> value
+  | Background_image, value -> value
+  | (Background_origin | Background_clip | Webkit_background_clip), value ->
+      value
+  | Aspect_ratio, value -> value
+  | Vertical_align, value -> value
+  | Font_family, value -> value
+  | (Background_position | Webkit_mask_position | Mask_position), value -> value
+  | (Background_repeat | Webkit_mask_repeat | Mask_repeat), value -> value
+  | Webkit_font_smoothing, value -> value
+  | Moz_osx_font_smoothing, value -> value
+  | Webkit_line_clamp, value -> value
+  | Webkit_box_orient, value -> value
+  | Moz_orient, value -> value
+  | Text_overflow, value -> value
+  | Text_wrap, value -> value
+  | Word_break, value -> value
+  | Overflow_wrap, value -> value
+  | Line_break, value -> value
+  | (Hyphens | Webkit_hyphens), value -> value
+  | Font_stretch, value -> value
+  | Font_optical_sizing, value -> value
+  | Font_kerning, value -> value
+  | Font_language_override, value -> value
+  | Font_synthesis_style, value -> value
+  | Font_synthesis_weight, value -> value
+  | Font_synthesis_small_caps, value -> value
+  | Font_synthesis_position, value -> value
+  | Font_variant_ligatures, value -> value
+  | Caps, value -> value
+  | Numeric, value -> value
+  | Font_variant_position, value -> value
+  | East_asian, value -> value
+  | Webkit_mask_composite, value -> value
+  | Webkit_mask_source_type, value -> value
+  | (Webkit_mask_clip | Webkit_mask_origin | Mask_clip | Mask_origin), value ->
+      value
+  | Mask_composite, value -> value
+  | Mask_mode, value -> value
+  | Mask_type, value -> value
+  | Scroll_snap_align, value -> value
+  | Scroll_snap_stop, value -> value
+  | Scroll_behavior, value -> value
+  | Field_sizing, value -> value
+  | Caption_side, value -> value
+  | Resize, value -> value
+  | Object_fit, value -> value
+  | Object_view_box, value -> value
+  | Color_scheme, value -> value
+  | (Print_color_adjust | Webkit_print_color_adjust), value -> value
+  | (Box_decoration_break | Webkit_box_decoration_break), value -> value
+  | Content, value -> value
+  | (Counter_reset | Counter_increment), value -> value
+  | Quotes, value -> value
+  | Touch_action, value -> value
+  | Clip, value -> value
+  | Clear, value -> value
+  | Float, value -> value
+  | Scale, value -> value
+  | (Fill | Stroke), value -> value
+  | Stroke_width, value -> value
+  | (Fill_rule | Clip_rule), value -> value
+  | Stroke_linecap, value -> value
+  | Stroke_linejoin, value -> value
+  | Stroke_miterlimit, value -> value
+  | Stroke_dashoffset, value -> value
+  | Stroke_dasharray, value -> value
+  | Paint_order, value -> value
+  | Vector_effect, value -> value
+  | Direction, value -> value
+  | Unicode_bidi, value -> value
+  | Writing_mode, value -> value
+  | Text_decoration_skip_ink, value -> value
+  | Animation_composition, value -> value
+  | Background_blend_mode, value -> value
+  | Overscroll_behavior, value -> value
+  | ( ( Overscroll_behavior_x | Overscroll_behavior_y
+      | Overscroll_behavior_block | Overscroll_behavior_inline ),
+      value ) ->
+      value
 
 (* CSS Values L4 sec. 10.10 ("Mathematical Expressions"): inside a math function
    ([calc], [min], [max], [clamp], [round], [mod], [rem], the trig family,
@@ -3596,6 +3984,9 @@ let normalize_property_value : type a.
   | Webkit_text_decoration -> normalize_text_decoration ~lossless value
   | Text_emphasis -> normalize_text_emphasis ~lossless value
   | Caret -> normalize_caret ~lossless value
+  | Interest_delay -> normalize_interest_delay value
+  | Interest_delay_start -> normalize_interest_delay value
+  | Interest_delay_end -> normalize_interest_delay value
   | Fill -> normalize_svg_paint ~lossless value
   | Stroke -> normalize_svg_paint ~lossless value
   | Scrollbar_color -> normalize_scrollbar_color ~lossless value
@@ -3841,15 +4232,15 @@ let pp_property_value : type a. (a property * a) Pp.t =
   | Row_gap -> pp pp_length
   | Width -> pp pp_length_percentage
   | Height -> pp pp_length_percentage
-  | Min_width -> pp pp_length_min_max
-  | Min_height -> pp pp_length_min_max
+  | Min_width -> pp pp_length_percentage
+  | Min_height -> pp pp_length_percentage
   | Max_width -> pp pp_length_percentage
   | Max_height -> pp pp_length_percentage
   | Inline_size -> pp pp_length_percentage
-  | Min_inline_size -> pp pp_length_min_max
+  | Min_inline_size -> pp pp_length_percentage
   | Max_inline_size -> pp pp_length_percentage
   | Block_size -> pp pp_length_percentage
-  | Min_block_size -> pp pp_length_min_max
+  | Min_block_size -> pp pp_length_percentage
   | Max_block_size -> pp pp_length_percentage
   | Font_size -> pp pp_font_size
   | Line_height -> pp pp_line_height